Keep BV validator cleanup below warning thresholds
The BV validator cluster was the safest remaining TypeScript maintenance slice because full catalog BV validation can prove parity after extraction. This moves large data/rule/setup/weapon-scan sections into focused helper modules and records the warning burn-down in the maintenance ledger.

Constraint: Preserve BattleMech BV parity across calculator, critical-slot scan, and ammo resolution.

Rejected: Tuning scanner thresholds | Would hide the maintenance debt instead of reducing it.

Confidence: high

Scope-risk: moderate

Directive: Keep future BV helper splits behind validate:bv and targeted maintenance scans.

Tested: npm.cmd run validate:bv; npm.cmd test -- --watchAll=false --runTestsByPath scripts/__tests__/validate-bv-fail-loud.test.ts scripts/__tests__/wave4-bv-report-output.test.ts --runInBand; npm.cmd run verify:qc:maintenance; npx.cmd tsc --noEmit --skipLibCheck; npx.cmd oxfmt --check changed files; git diff --cached --check

Not-tested: Full verify:qc:roadmap was not rerun for this narrow maintenance slice.
